pyproject.toml
[project]
name = "weaviate-pydantic-ai-workshop"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "numpy>=2.2.6",
    "pydantic-ai>=1.0.15",
    "pymupdf>=1.26.4",
    "weaviate-agents>=1.0.1",
    "weaviate-client>=4.17.0",
]
